requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.7.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Flask==2.2.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.26.5
OpenCC==1.1.6
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.28.2
six==1.16.0
sniffio==1.3.1
telegram==0.0.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzlocal==2.1
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
